[projectfirma/#2155] Project Running Balance: use Larry's Special as source Renaming "AllContract" classes => "ObligationsAndExpenditures"
