+++ Source/ProjectFirma.Web/Views/Project/Detail.cshtml
@@ -33,14 +33,13 @@
 @using ProjectFirma.Web.Common
 @using ProjectFirma.Web.Models
 @using ProjectFirma.Web.Views.ActionItem
-@using ProjectFirma.Web.Views.ProjectFunding
 @using ProjectFirma.Web.Views.ProjectAgreement
 @using ProjectFirma.Web.Views.Shared.PerformanceMeasureControls
 @using ProjectFirma.Web.Views.Shared.ProjectAttachment
 @using ProjectFirma.Web.Views.Shared.ProjectContact
 @using ProjectFirma.Web.Views.Shared.ProjectOrganization
 @using ProjectFirma.Web.Views.Shared.ProjectRunningBalance
-@using ProjectFirma.Web.Views.Shared.ProjectRunningBalanceAllContract
+@using ProjectFirma.Web.Views.Shared.ProjectRunningBalanceObligationsAndExpenditures
 @using ProjectFirma.Web.Views.TechnicalAssistanceRequest
 @using ProjectCustomAttributes = ProjectFirma.Web.Views.Shared.ProjectControls.DisplayProjectCustomAttributes
 @using ProjectTimeline = ProjectFirma.Web.Views.Shared.ProjectTimeline.ProjectTimelineDisplay
@@ -430,11 +429,11 @@
             <div class="panel panelFirma">
                <div class="panel-heading panelTitle">
                   <h3 class="project-detail-heading">
-                     Project Running Balance - All Contracts
+                     Project Running Balance - Obligations and Expenditures
                   </h3>
                </div>
                <div class="panel-body">
-                  @{ ProjectRunningBalanceAllContract.RenderPartialView(Html, ViewDataTyped.ProjectRunningBalanceAllContractViewData); }
+                  @{ ProjectRunningBalanceObligationsAndExpenditures.RenderPartialView(Html, ViewDataTyped.ProjectRunningBalanceObligationsAndExpendituresViewData); }
                </div>
             </div>
