+++ Source/ProjectFirma.Web/Views/Project/DetailViewData.cs
@@ -41,7 +41,7 @@ using ProjectFirma.Web.Views.Obligation;
 using ProjectFirma.Web.Views.Shared.ProjectAttachment;
 using ProjectFirma.Web.Views.ProjectFunding;
 using ProjectFirma.Web.Views.Shared.ProjectRunningBalance;
-using ProjectFirma.Web.Views.Shared.ProjectRunningBalanceAllContract;
+using ProjectFirma.Web.Views.Shared.ProjectRunningBalanceObligationsAndExpenditures;
 
 namespace ProjectFirma.Web.Views.Project
 {
@@ -146,7 +146,7 @@ namespace ProjectFirma.Web.Views.Project
 
         //Project Running Balance
         public ProjectRunningBalanceViewData ProjectRunningBalanceViewData { get; set; }
-        public ProjectRunningBalanceAllContractViewData ProjectRunningBalanceAllContractViewData { get; set; }
+        public ProjectRunningBalanceObligationsAndExpendituresViewData ProjectRunningBalanceObligationsAndExpendituresViewData { get; set; }
 
         public string UpdateStatusUrl { get; set; }
         public DetailViewData(FirmaSession currentFirmaSession, ProjectFirmaModels.Models.Project project,
@@ -187,7 +187,7 @@ namespace ProjectFirma.Web.Views.Project
             ActionItemsDisplayViewData actionItemsDisplayViewData,
             bool userCanViewActionItems,
             ProjectRunningBalanceViewData projectRunningBalanceViewData,
-            ProjectRunningBalanceAllContractViewData projectRunningBalanceAllContractViewData)
+            ProjectRunningBalanceObligationsAndExpendituresViewData projectRunningBalanceObligationsAndExpendituresViewData)
             : base(currentFirmaSession, project)
         {
             PageTitle = project.GetDisplayName();
@@ -449,7 +449,7 @@ namespace ProjectFirma.Web.Views.Project
 
             // Project Running Balance
             ProjectRunningBalanceViewData = projectRunningBalanceViewData;
-            ProjectRunningBalanceAllContractViewData = projectRunningBalanceAllContractViewData;
+            ProjectRunningBalanceObligationsAndExpendituresViewData = projectRunningBalanceObligationsAndExpendituresViewData;
         }
     }
 }
